flask-blog-app/requirements.txt
flask
flask_mysqldb
wtforms
passlib.hash
functools